Moved database connection into dataManagement package and split questions into designated classes I am doing the first 3.
