+++ Creative_Chipmunks_Assignment09/main_package/main.py
@@ -17,18 +17,3 @@
 if __name__ == "__main__":
     pass
 
-
-import pyodbc
-def connect_to_database(self):
-        """
-        Connect to our SQL Server instance
-        @return connection object
-        """
-        conn = pyodbc.connect(
-            'Driver={SQL Server};'
-            'Server=lcb-sql.uccob.uc.edu\\nicholdw;'
-            'Database=GroceryStoreSimulator;'
-            'uid=IS4010Login;'
-            'pwd=P@ssword2;'
-        )
-        return conn
